+++ soft_render/src/SR_Texture.cpp
@@ -254,7 +254,7 @@ int SR_Texture::init(SR_ColorDataType type, uint16_t w, uint16_t h, uint16_t d) 
     mDepthf        = (float)d;
     mType          = type;
     mBytesPerTexel = (uint32_t)bpt;
-    mNumChannels = sr_elements_per_color(type);
+    mNumChannels = (uint32_t)sr_elements_per_color(type);
     mTexels        = pData;
 
     return 0;
